Se agrega toda la capa de documentación con Swagger

+++ src/main/java/com/oo2/grupo17/dtos/ClienteRegistroDto.java
@@ -33,7 +33,7 @@ public class ClienteRegistroDto {
     @NotNull(message = "El móvil es obligatorio")
     @Min(value = 1000000000, message = "El móvil debe tener al menos 10 dígitos")
     @Max(value = 9999999999L, message = "El móvil no debe superar los 10 dígitos")
-    private int movil;
+    private long movil;
 
     @Min(value = 1000000, message = "El teléfono debe tener al menos 7 dígitos")
     @Max(value = 9999999999L, message = "El teléfono no debe superar los 10 dígitos")
